+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/table/DefaultFilterTableItems.java
@@ -497,7 +497,6 @@ final class DefaultFilterTableItems<R, C> implements FilterTableModelItems<R> {
 				if (visiblePredicate.test(item)) {
 					items.set(index, item);
 					tableModel.fireTableRowsUpdated(index, index);
-					visible.sort();
 
 					return true;
 				}